pyproject.toml
[project]
name = "vault"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0.3",
]
